[logic][m]: get_package moved to logic - refs #401
